BaseTile now extends render so that tiles can be drawn at iso coords
I imagine other things will use the Renderable.render to draw stuff that doesn't use iso coords

+++ src/main/java/com/knightlore/client/render/Renderable.java
@@ -8,7 +8,8 @@ public abstract class Renderable {
   protected RenderModel model;
   protected Texture texture;
 
-  public void render(int x, int y, ShaderProgram shaderProgram, Matrix4f world, Matrix4f camera) {
+  public void render(
+      float x, float y, ShaderProgram shaderProgram, Matrix4f world, Matrix4f camera) {
     shaderProgram.bind();
     texture.bind(0);
